changed folder structure in components
