+++ pages/signin/index.js
@@ -10,6 +10,7 @@ import GoogleOAuth from "../../components/elements/GoogleOAuth";
 import { UserContext } from "../../components/context";
 import { useGoogleReCaptcha } from "react-google-recaptcha-v3";
 import SEO from "../../components/utils/SEO";
+import Footer from "../../components/elements/Footer";
 export default function SignIn() {
     const { executeRecaptcha } = useGoogleReCaptcha();
     const router = useRouter();
@@ -131,6 +132,7 @@ export default function SignIn() {
                     </form>
                 </div>
             </div>
+            <Footer />
         </>
     );
 }
